pyproject.toml
[project]
name = "591-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler==3.11.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "exceptiongroup==1.2.2",
    "h11==0.14.0",
    "idna==3.10",
    "outcome==1.3.0.post0",
    "pysocks==1.7.1",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "selenium==4.31.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "trio==0.29.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.13.2",
    "tzlocal==5.3.1",
    "urllib3==2.4.0",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
]
